ref(tests): Extract simulate_impersonation into shared test helper
Move duplicated _simulate_impersonation method from 7 test files into
a shared helper at testutils/helpers/impersonation.py. The helper
takes the impersonator user as a parameter instead of reading it from
self, making it a simple standalone context manager.
