+++ tests/sentry/api/endpoints/test_api_tokens.py
@@ -1,6 +1,4 @@
 from collections.abc import Generator
-from typing import Any
-from unittest.mock import patch
 
 from django.urls import reverse
 from pytest import fixture
@@ -8,6 +6,7 @@ from rest_framework import status
 
 from sentry.models.apitoken import ApiToken
 from sentry.testutils.cases import APITestCase
+from sentry.testutils.helpers.impersonation import simulate_impersonation
 from sentry.testutils.helpers.options import override_options
 from sentry.testutils.silo import control_silo_test
 
@@ -325,22 +324,9 @@ class ApiTokensImpersonationTest(APITestCase):
         self.impersonator = self.create_user(is_superuser=True)
         self.target_user = self.create_user()
 
-    def _simulate_impersonation(self) -> Any:
-        """Patch Endpoint.initialize_request to set actual_user, simulating ViewAs middleware."""
-        from sentry.api.base import Endpoint
-
-        original = Endpoint.initialize_request
-
-        def patched(endpoint_self: Any, request: Any, *args: Any, **kwargs: Any) -> Any:
-            drf_request = original(endpoint_self, request, *args, **kwargs)
-            drf_request.actual_user = self.impersonator  # type: ignore[attr-defined]
-            return drf_request
-
-        return patch.object(Endpoint, "initialize_request", patched)
-
     def test_impersonated_post_blocked(self) -> None:
         self.login_as(self.target_user)
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.post(self.url, data={"scopes": ["event:read"]})
         assert response.status_code == status.HTTP_403_FORBIDDEN
         assert not ApiToken.objects.filter(user=self.target_user).exists()
@@ -348,7 +334,7 @@ class ApiTokensImpersonationTest(APITestCase):
     def test_impersonated_delete_blocked(self) -> None:
         token = ApiToken.objects.create(user=self.target_user)
         self.login_as(self.target_user)
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.delete(self.url, data={"tokenId": token.id})
         assert response.status_code == status.HTTP_403_FORBIDDEN
         assert ApiToken.objects.filter(id=token.id).exists()
@@ -356,7 +342,7 @@ class ApiTokensImpersonationTest(APITestCase):
     def test_impersonated_get_allowed(self) -> None:
         ApiToken.objects.create(user=self.target_user)
         self.login_as(self.target_user)
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.get(self.url)
         assert response.status_code == status.HTTP_200_OK
