+++ tests/sentry/api/endpoints/test_organization_auth_tokens.py
@@ -1,6 +1,3 @@
-from typing import Any
-from unittest.mock import patch
-
 from django.urls import reverse
 from rest_framework import status
 
@@ -8,6 +5,7 @@ from sentry import options
 from sentry.models.apitoken import ApiToken
 from sentry.models.orgauthtoken import OrgAuthToken
 from sentry.testutils.cases import APITestCase, PermissionTestCase
+from sentry.testutils.helpers.impersonation import simulate_impersonation
 from sentry.testutils.silo import control_silo_test, create_test_regions
 from sentry.types.region import get_cell_by_name
 from sentry.utils.security.orgauthtoken_token import parse_token
@@ -238,22 +236,10 @@ class OrganizationAuthTokensImpersonationTest(APITestCase):
         super().setUp()
         self.impersonator = self.create_user(is_superuser=True)
 
-    def _simulate_impersonation(self) -> Any:
-        from sentry.api.base import Endpoint
-
-        original = Endpoint.initialize_request
-
-        def patched(endpoint_self: Any, request: Any, *args: Any, **kwargs: Any) -> Any:
-            drf_request = original(endpoint_self, request, *args, **kwargs)
-            drf_request.actual_user = self.impersonator  # type: ignore[attr-defined]
-            return drf_request
-
-        return patch.object(Endpoint, "initialize_request", patched)
-
     def test_impersonated_post_blocked(self) -> None:
         self.login_as(self.user)
         url = reverse("sentry-api-0-org-auth-tokens", args=[self.organization.slug])
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.post(url, data={"name": "test token"})
         assert response.status_code == status.HTTP_403_FORBIDDEN
         assert not OrgAuthToken.objects.filter(organization_id=self.organization.id).exists()
@@ -261,6 +247,6 @@ class OrganizationAuthTokensImpersonationTest(APITestCase):
     def test_impersonated_get_allowed(self) -> None:
         self.login_as(self.user)
         url = reverse("sentry-api-0-org-auth-tokens", args=[self.organization.slug])
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.get(url)
         assert response.status_code == status.HTTP_200_OK
