+++ tests/sentry/releases/endpoints/test_project_releases_token.py
@@ -1,10 +1,8 @@
-from typing import Any
-from unittest.mock import patch
-
 from django.urls import reverse
 
 from sentry.models.options.project_option import ProjectOption
 from sentry.testutils.cases import APITestCase
+from sentry.testutils.helpers.impersonation import simulate_impersonation
 from sentry.types.region import get_local_locality
 
 
@@ -95,18 +93,6 @@ class ReleaseTokenImpersonationTest(APITestCase):
         super().setUp()
         self.impersonator = self.create_user(is_superuser=True)
 
-    def _simulate_impersonation(self) -> Any:
-        from sentry.api.base import Endpoint
-
-        original = Endpoint.initialize_request
-
-        def patched(endpoint_self: Any, request: Any, *args: Any, **kwargs: Any) -> Any:
-            drf_request = original(endpoint_self, request, *args, **kwargs)
-            drf_request.actual_user = self.impersonator  # type: ignore[attr-defined]
-            return drf_request
-
-        return patch.object(Endpoint, "initialize_request", patched)
-
     def test_impersonated_post_blocked(self) -> None:
         project = self.create_project(name="foo")
         token = "abcdefghijklmnop"
@@ -120,7 +106,7 @@ class ReleaseTokenImpersonationTest(APITestCase):
             },
         )
         self.login_as(user=self.user)
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.post(url, {"project": project.slug})
         assert response.status_code == 403
         assert ProjectOption.objects.get_value(project, "sentry:release-token") == token
@@ -137,7 +123,7 @@ class ReleaseTokenImpersonationTest(APITestCase):
             },
         )
         self.login_as(user=self.user)
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.get(url)
         assert response.status_code == 200
 
@@ -152,7 +138,7 @@ class ReleaseTokenImpersonationTest(APITestCase):
             },
         )
         self.login_as(user=self.user)
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.client.get(url)
         assert response.status_code == 404
         assert ProjectOption.objects.get_value(project, "sentry:release-token") is None
