+++ tests/sentry/sentry_apps/api/endpoints/test_sentry_internal_app_token_details.py
@@ -1,6 +1,3 @@
-from typing import Any
-from unittest.mock import patch
-
 from django.test import override_settings
 from rest_framework import status
 
@@ -8,6 +5,7 @@ from sentry import audit_log
 from sentry.models.apitoken import ApiToken
 from sentry.testutils.asserts import assert_org_audit_log_exists
 from sentry.testutils.cases import APITestCase
+from sentry.testutils.helpers.impersonation import simulate_impersonation
 from sentry.testutils.helpers.options import override_options
 from sentry.testutils.silo import control_silo_test
 
@@ -154,21 +152,9 @@ class SentryInternalAppTokenDetailsImpersonationTest(APITestCase):
         )
         self.impersonator = self.create_user(is_superuser=True)
 
-    def _simulate_impersonation(self) -> Any:
-        from sentry.api.base import Endpoint
-
-        original = Endpoint.initialize_request
-
-        def patched(endpoint_self: Any, request: Any, *args: Any, **kwargs: Any) -> Any:
-            drf_request = original(endpoint_self, request, *args, **kwargs)
-            drf_request.actual_user = self.impersonator  # type: ignore[attr-defined]
-            return drf_request
-
-        return patch.object(Endpoint, "initialize_request", patched)
-
     def test_impersonated_delete_blocked(self) -> None:
         self.login_as(self.user)
-        with self._simulate_impersonation():
+        with simulate_impersonation(self.impersonator):
             response = self.get_error_response(
                 self.internal_sentry_app.slug,
                 self.api_token.id,
